+++ rpc-server/src/modules/blocks/methods.rs
@@ -4,7 +4,8 @@ use near_primitives::views::StateChangeValueView;
 
 use crate::config::ServerContext;
 use crate::modules::blocks::utils::{
-    check_block_height, fetch_block_from_cache_or_get, fetch_chunk_from_s3, is_matching_change,
+    check_block_height, fetch_block_from_cache_or_get, fetch_chunk_from_fastnear,
+    is_matching_change,
 };
 
 /// `block` rpc method implementation
@@ -260,7 +261,13 @@ async fn changes_in_block_call(
         }
     })?;
 
-    let result = fetch_changes_in_block(&data, cache_block, &params.block_reference).await;
+    let result = fetch_changes_in_block(
+        &data,
+        cache_block,
+        &params.block_reference,
+        "EXPERIMENTAL_changes_in_block",
+    )
+    .await;
     #[cfg(feature = "shadow-data-consistency")]
     {
         if let near_primitives::types::BlockReference::Finality(_) = params.block_reference {
@@ -305,6 +312,7 @@ async fn changes_in_block_by_type_call(
         cache_block,
         &params.state_changes_request,
         &params.block_reference,
+        "EXPERIMENTAL_changes",
     )
     .await;
 
@@ -366,6 +374,9 @@ pub async fn fetch_block(
             return match finality {
                 near_primitives::types::Finality::Final
                 | near_primitives::types::Finality::DoomSlug => {
+                    crate::metrics::REQUESTS_BLOCKS_COUNTERS
+                        .with_label_values(&[method_name, "cache"])
+                        .inc();
                     let block_view = data.blocks_info_by_finality.final_block_view().await;
                     Ok(near_jsonrpc::primitives::types::blocks::RpcBlockResponse { block_view })
                 }
@@ -378,6 +389,9 @@ pub async fn fetch_block(
                             },
                         )
                     } else {
+                        crate::metrics::REQUESTS_BLOCKS_COUNTERS
+                            .with_label_values(&[method_name, "cache"])
+                            .inc();
                         let block_view = data.blocks_info_by_finality.optimistic_block_view().await;
                         Ok(
                             near_jsonrpc::primitives::types::blocks::RpcBlockResponse {
@@ -401,6 +415,9 @@ pub async fn fetch_block(
             .await
             .block_height
     {
+        crate::metrics::REQUESTS_BLOCKS_COUNTERS
+            .with_label_values(&[method_name, "cache"])
+            .inc();
         data.blocks_info_by_finality.final_block_view().await
     } else if block_height
         == data
@@ -409,16 +426,21 @@ pub async fn fetch_block(
             .await
             .block_height
     {
+        crate::metrics::REQUESTS_BLOCKS_COUNTERS
+            .with_label_values(&[method_name, "cache"])
+            .inc();
         data.blocks_info_by_finality.optimistic_block_view().await
     } else {
-        near_lake_framework::s3::fetchers::fetch_block(
-            &data.s3_client,
-            &data.s3_bucket_name,
+        crate::metrics::REQUESTS_BLOCKS_COUNTERS
+            .with_label_values(&[method_name, "lake"])
+            .inc();
+        near_lake_framework::fastnear::fetchers::fetch_block_or_retry(
+            &data.fastnear_client,
             block_height,
         )
         .await
         .map_err(|err| {
-            tracing::error!("Failed to fetch block from S3: {}", err);
+            tracing::error!("Failed to fetch block from fastnear: {}", err);
             near_jsonrpc::primitives::types::blocks::RpcBlockError::UnknownBlock {
                 error_message: format!("BLOCK HEIGHT: {:?}", block_height),
             }
@@ -484,13 +506,11 @@ pub async fn fetch_chunk(
             )
             .map(|block_height_shard_id| (block_height_shard_id.0, block_height_shard_id.1))?,
     };
-    let chunk_view = fetch_chunk_from_s3(
-        &data.s3_client,
-        &data.s3_bucket_name,
-        block_height,
-        shard_id.into(),
-    )
-    .await?;
+    crate::metrics::REQUESTS_BLOCKS_COUNTERS
+        .with_label_values(&[method_name, "lake"])
+        .inc();
+    let chunk_view =
+        fetch_chunk_from_fastnear(&data.fastnear_client, block_height, shard_id.into()).await?;
     // increase block category metrics
     crate::metrics::increase_request_category_metrics(
         data,
@@ -510,11 +530,12 @@ async fn fetch_changes_in_block(
     data: &Data<ServerContext>,
     cache_block: crate::modules::blocks::CacheBlock,
     block_reference: &near_primitives::types::BlockReference,
+    method_name: &str,
 ) -> Result<
     near_jsonrpc::primitives::types::changes::RpcStateChangesInBlockByTypeResponse,
     near_jsonrpc::primitives::types::changes::RpcStateChangesError,
 > {
-    let trie_keys = fetch_state_changes(data, cache_block, block_reference)
+    let trie_keys = fetch_state_changes(data, cache_block, block_reference, method_name)
         .await
         .map_err(|err| {
             near_jsonrpc::primitives::types::changes::RpcStateChangesError::UnknownBlock {
@@ -598,11 +619,12 @@ async fn fetch_changes_in_block_by_type(
     cache_block: crate::modules::blocks::CacheBlock,
     state_changes_request: &near_primitives::views::StateChangesRequestView,
     block_reference: &near_primitives::types::BlockReference,
+    method_name: &str,
 ) -> Result<
     near_jsonrpc::primitives::types::changes::RpcStateChangesInBlockResponse,
     near_jsonrpc::primitives::types::changes::RpcStateChangesError,
 > {
-    let changes = fetch_state_changes(data, cache_block, block_reference)
+    let changes = fetch_state_changes(data, cache_block, block_reference, method_name)
         .await
         .map_err(|err| {
             near_jsonrpc::primitives::types::changes::RpcStateChangesError::UnknownBlock {
@@ -629,6 +651,7 @@ async fn fetch_state_changes(
     data: &Data<ServerContext>,
     cache_block: crate::modules::blocks::CacheBlock,
     block_reference: &near_primitives::types::BlockReference,
+    method_name: &str,
 ) -> anyhow::Result<near_primitives::views::StateChangesView> {
     if let near_primitives::types::BlockReference::Finality(finality) = block_reference {
         match finality {
@@ -650,7 +673,7 @@ async fn fetch_state_changes(
             }
         }
     } else {
-        Ok(fetch_shards_by_cache_block(data, cache_block)
+        Ok(fetch_shards_by_cache_block(data, cache_block, method_name)
             .await?
             .into_iter()
             .flat_map(|shard| shard.state_changes)
@@ -663,28 +686,21 @@ async fn fetch_state_changes(
 async fn fetch_shards_by_cache_block(
     data: &Data<ServerContext>,
     cache_block: crate::modules::blocks::CacheBlock,
+    method_name: &str,
 ) -> anyhow::Result<Vec<near_indexer_primitives::IndexerShard>> {
-    let fetch_shards_futures = (0..cache_block.chunks_included)
-        .collect::<Vec<u64>>()
-        .into_iter()
-        .map(|shard_id| {
-            near_lake_framework::s3::fetchers::fetch_shard(
-                &data.s3_client,
-                &data.s3_bucket_name,
-                cache_block.block_height,
-                shard_id,
-            )
-        });
-
-    futures::future::join_all(fetch_shards_futures)
-        .await
-        .into_iter()
-        .collect::<Result<_, _>>()
-        .map_err(|err| {
-            anyhow::anyhow!(
-                "Failed to fetch shards for block {} with error: {}",
-                cache_block.block_height,
-                err
-            )
-        })
+    crate::metrics::REQUESTS_BLOCKS_COUNTERS
+        .with_label_values(&[method_name, "lake"])
+        .inc();
+    match near_lake_framework::fastnear::fetchers::fetch_streamer_message(
+        &data.fastnear_client,
+        cache_block.block_height,
+    )
+    .await
+    {
+        Some(streamer_message) => Ok(streamer_message.shards),
+        None => Err(anyhow::anyhow!(
+            "Failed to fetch shards for block {}",
+            cache_block.block_height,
+        )),
+    }
 }